--- a/Makefile
+++ b/Makefile
@@ -151,6 +151,10 @@ OSDK_CRATES := \
 	kernel/libs/aster-util \
 	kernel/libs/aster-bigtcp
 
+# OSDK dependencies
+OSDK_SRC_FILES := \
+	$(shell find osdk/Cargo.toml osdk/Cargo.lock osdk/src -type f)
+
 .PHONY: all
 all: build
 
@@ -163,10 +167,9 @@ install_osdk:
 	@# dependencies to `crates.io`.
 	@OSDK_LOCAL_DEV=1 cargo install cargo-osdk --path osdk
 
-# This will install OSDK if it is not already installed
-# To update OSDK, we need to run `install_osdk` manually
-$(CARGO_OSDK):
-	@make --no-print-directory install_osdk
+# This will install and update OSDK automatically
+$(CARGO_OSDK): $(OSDK_SRC_FILES)
+	@$(MAKE) --no-print-directory install_osdk
 
 .PHONY: check_osdk
 check_osdk:
@@ -180,7 +183,7 @@ test_osdk:
 
 .PHONY: initramfs
 initramfs:
-	@make --no-print-directory -C test
+	@$(MAKE) --no-print-directory -C test
 
 .PHONY: build
 build: initramfs $(CARGO_OSDK)
@@ -213,7 +216,7 @@ gdb_server: initramfs $(CARGO_OSDK)
 	@cargo osdk run $(CARGO_OSDK_ARGS) --gdb-server wait-client,vscode,addr=:$(GDB_TCP_PORT)
 
 .PHONY: gdb_client
-gdb_client: $(CARGO_OSDK)
+gdb_client: initramfs $(CARGO_OSDK)
 	@cargo osdk debug $(CARGO_OSDK_ARGS) --remote :$(GDB_TCP_PORT)
 
 .PHONY: profile_server
@@ -221,7 +224,7 @@ profile_server: initramfs $(CARGO_OSDK)
 	@cargo osdk run $(CARGO_OSDK_ARGS) --gdb-server addr=:$(GDB_TCP_PORT)
 
 .PHONY: profile_client
-profile_client: $(CARGO_OSDK)
+profile_client: initramfs $(CARGO_OSDK)
 	@cargo osdk profile $(CARGO_OSDK_ARGS) --remote :$(GDB_TCP_PORT) \
 		--samples $(GDB_PROFILE_COUNT) --interval $(GDB_PROFILE_INTERVAL) --format $(GDB_PROFILE_FORMAT)
 
@@ -255,7 +258,7 @@ docs: $(CARGO_OSDK)
 .PHONY: format
 format:
 	@./tools/format_all.sh
-	@make --no-print-directory -C test format
+	@$(MAKE) --no-print-directory -C test format
 
 .PHONY: check
 check: initramfs $(CARGO_OSDK)
@@ -277,7 +280,7 @@ check: initramfs $(CARGO_OSDK)
 		echo "Checking $$dir"; \
 		(cd $$dir && cargo osdk clippy -- -- -D warnings) || exit 1; \
 	done
-	@make --no-print-directory -C test check
+	@$(MAKE) --no-print-directory -C test check
 	@typos
 
 .PHONY: clean
@@ -289,6 +292,6 @@ clean:
 	@echo "Cleaning up documentation target files"
 	@cd docs && mdbook clean
 	@echo "Cleaning up test target files"
-	@make --no-print-directory -C test clean
+	@$(MAKE) --no-print-directory -C test clean
 	@echo "Uninstalling OSDK"
 	@rm -f $(CARGO_OSDK)