--- a/osdk/src/base_crate/mod.rs
+++ b/osdk/src/base_crate/mod.rs
@@ -5,11 +5,44 @@
 
 use std::{
     fs,
+    io::{Read, Result},
     path::{Path, PathBuf},
     str::FromStr,
 };
 
 use crate::util::get_cargo_metadata;
+
+/// Compares two files byte-by-byte to check if they are identical.
+/// Returns `Ok(true)` if files are identical, `Ok(false)` if they are different, or `Err` if any I/O operation fails.
+fn are_files_identical(file1: &PathBuf, file2: &PathBuf) -> Result<bool> {
+    // Check file size first
+    let metadata1 = fs::metadata(file1)?;
+    let metadata2 = fs::metadata(file2)?;
+
+    if metadata1.len() != metadata2.len() {
+        return Ok(false); // Different sizes, not identical
+    }
+
+    // Compare file contents byte-by-byte
+    let mut file1 = fs::File::open(file1)?;
+    let mut file2 = fs::File::open(file2)?;
+
+    let mut buffer1 = [0u8; 4096];
+    let mut buffer2 = [0u8; 4096];
+
+    loop {
+        let bytes_read1 = file1.read(&mut buffer1)?;
+        let bytes_read2 = file2.read(&mut buffer2)?;
+
+        if bytes_read1 != bytes_read2 || buffer1[..bytes_read1] != buffer2[..bytes_read1] {
+            return Ok(false); // Files are different
+        }
+
+        if bytes_read1 == 0 {
+            return Ok(true); // End of both files, identical
+        }
+    }
+}
 
 /// Create a new base crate that will be built by cargo.
 ///
@@ -17,6 +50,43 @@ use crate::util::get_cargo_metadata;
 /// `link_unit_test_runner` is set to true, the base crate will also depend on
 /// the `ostd-test-runner` crate.
 pub fn new_base_crate(
+    base_crate_path: impl AsRef<Path>,
+    dep_crate_name: &str,
+    dep_crate_path: impl AsRef<Path>,
+    link_unit_test_runner: bool,
+) {
+    // Check if the existing crate base is reusable. Crate bases for ktest are never reusable.
+    if !base_crate_path.as_ref().ends_with("test-base") && base_crate_path.as_ref().exists() {
+        let base_crate_tmp_path = base_crate_path.as_ref().join("tmp");
+        do_new_base_crate(
+            &base_crate_tmp_path,
+            dep_crate_name,
+            &dep_crate_path,
+            link_unit_test_runner,
+        );
+        let cargo_result = are_files_identical(
+            &base_crate_path.as_ref().join("Cargo.toml"),
+            &base_crate_tmp_path.join("Cargo.toml"),
+        );
+        let main_rs_result = are_files_identical(
+            &base_crate_path.as_ref().join("src").join("main.rs"),
+            &base_crate_tmp_path.join("src").join("main.rs"),
+        );
+        std::fs::remove_dir_all(&base_crate_tmp_path).unwrap();
+        if cargo_result.is_ok_and(|res| res) && main_rs_result.is_ok_and(|res| res) {
+            info!("Reusing existing base crate");
+            return;
+        }
+    }
+    do_new_base_crate(
+        base_crate_path,
+        dep_crate_name,
+        dep_crate_path,
+        link_unit_test_runner,
+    );
+}
+
+fn do_new_base_crate(
     base_crate_path: impl AsRef<Path>,
     dep_crate_name: &str,
     dep_crate_path: impl AsRef<Path>,