--- a/osdk/src/bundle/file.rs
+++ b/osdk/src/bundle/file.rs
@@ -53,7 +53,7 @@ impl Initramfs {
         }
     }
 
-    /// Move the initramfs to the `base` directory and convert the path to a relative path.
+    /// Copy the initramfs to the `base` directory and convert the path to a relative path.
     pub fn copy_to(self, base: impl AsRef<Path>) -> Self {
         let name = self.path.file_name().unwrap();
         let dest = base.as_ref().join(name);