--- a/osdk/src/bundle/file.rs
+++ b/osdk/src/bundle/file.rs
@@ -1,36 +1,39 @@
 // SPDX-License-Identifier: MPL-2.0
 
 use std::{
-    fs, io,
+    os::unix::fs::MetadataExt,
     path::{Path, PathBuf},
+    time::SystemTime,
 };
-
-use sha2::{Digest, Sha256};
 
 use crate::util::hard_link_or_copy;
 
-/// A trait for files in a bundle. The file in a bundle should have it's digest and be validatable.
+/// A trait for files in a bundle. The file in a bundle should have its modified time and be validatable.
 pub trait BundleFile {
     fn path(&self) -> &PathBuf;
 
-    fn sha256sum(&self) -> &String;
+    fn modified_time(&self) -> &SystemTime;
 
-    fn calculate_sha256sum(&self) -> String {
-        let mut file = fs::File::open(self.path()).unwrap();
-        let mut hasher = Sha256::new();
-        let _n = io::copy(&mut file, &mut hasher).unwrap();
-        format!("{:x}", hasher.finalize())
+    fn size(&self) -> &u64;
+
+    fn get_modified_time(&self) -> SystemTime {
+        self.path().metadata().unwrap().modified().unwrap()
+    }
+
+    fn get_size(&self) -> u64 {
+        self.path().metadata().unwrap().size()
     }
 
     fn validate(&self) -> bool {
-        self.sha256sum() == &self.calculate_sha256sum()
+        self.size() == &self.get_size() && self.modified_time() >= &self.get_modified_time()
     }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Initramfs {
     path: PathBuf,
-    sha256sum: String,
+    modified_time: SystemTime,
+    size: u64,
 }
 
 impl BundleFile for Initramfs {
@@ -38,8 +41,12 @@ impl BundleFile for Initramfs {
         &self.path
     }
 
-    fn sha256sum(&self) -> &String {
-        &self.sha256sum
+    fn modified_time(&self) -> &SystemTime {
+        &self.modified_time
+    }
+
+    fn size(&self) -> &u64 {
+        &self.size
     }
 }
 
@@ -47,10 +54,12 @@ impl Initramfs {
     pub fn new(path: impl AsRef<Path>) -> Self {
         let created = Self {
             path: path.as_ref().to_path_buf(),
-            sha256sum: String::new(),
+            modified_time: SystemTime::UNIX_EPOCH,
+            size: 0,
         };
         Self {
-            sha256sum: created.calculate_sha256sum(),
+            modified_time: created.get_modified_time(),
+            size: created.get_size(),
             ..created
         }
     }
@@ -59,9 +68,10 @@ impl Initramfs {
     pub fn copy_to(self, base: impl AsRef<Path>) -> Self {
         let name = self.path.file_name().unwrap();
         let dest = base.as_ref().join(name);
-        hard_link_or_copy(&self.path, dest).unwrap();
+        hard_link_or_copy(&self.path, &dest).unwrap();
         Self {
             path: PathBuf::from(name),
+            modified_time: dest.metadata().unwrap().modified().unwrap(),
             ..self
         }
     }