--- a/osdk/src/bundle/vm_image.rs
+++ b/osdk/src/bundle/vm_image.rs
@@ -1,6 +1,10 @@
 // SPDX-License-Identifier: MPL-2.0
 
-use std::path::{Path, PathBuf};
+use std::{
+    os::unix::fs::MetadataExt,
+    path::{Path, PathBuf},
+    time::SystemTime,
+};
 
 use crate::util::hard_link_or_copy;
 
@@ -11,7 +15,8 @@ pub struct AsterVmImage {
     path: PathBuf,
     typ: AsterVmImageType,
     aster_version: String,
-    sha256sum: String,
+    modified_time: SystemTime,
+    size: u64,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -35,8 +40,12 @@ impl BundleFile for AsterVmImage {
         &self.path
     }
 
-    fn sha256sum(&self) -> &String {
-        &self.sha256sum
+    fn modified_time(&self) -> &SystemTime {
+        &self.modified_time
+    }
+
+    fn size(&self) -> &u64 {
+        &self.size
     }
 }
 
@@ -46,10 +55,12 @@ impl AsterVmImage {
             path: path.as_ref().to_path_buf(),
             typ,
             aster_version,
-            sha256sum: String::new(),
+            modified_time: SystemTime::UNIX_EPOCH,
+            size: 0,
         };
         Self {
-            sha256sum: created.calculate_sha256sum(),
+            modified_time: created.get_modified_time(),
+            size: created.get_size(),
             ..created
         }
     }
@@ -62,12 +73,14 @@ impl AsterVmImage {
     pub fn copy_to(self, base: impl AsRef<Path>) -> Self {
         let file_name = self.path.file_name().unwrap();
         let copied_path = base.as_ref().join(file_name);
-        hard_link_or_copy(&self.path, copied_path).unwrap();
+        hard_link_or_copy(&self.path, &copied_path).unwrap();
+        let copied_metadata = copied_path.metadata().unwrap();
         Self {
             path: PathBuf::from(file_name),
             typ: self.typ,
             aster_version: self.aster_version,
-            sha256sum: self.sha256sum,
+            modified_time: copied_metadata.modified().unwrap(),
+            size: copied_metadata.size(),
         }
     }