--- a/osdk/src/bundle/vm_image.rs
+++ b/osdk/src/bundle/vm_image.rs
@@ -59,8 +59,8 @@ impl AsterVmImage {
         &self.typ
     }
 
-    /// Move the binary to the `base` directory and convert the path to a relative path.
-    pub fn move_to(self, base: impl AsRef<Path>) -> Self {
+    /// Copy the binary to the `base` directory and convert the path to a relative path.
+    pub fn copy_to(self, base: impl AsRef<Path>) -> Self {
         let file_name = self.path.file_name().unwrap();
         let copied_path = base.as_ref().join(file_name);
         fs::copy(&self.path, copied_path).unwrap();