This PR fixes bugs in OSDK that caused base crates, bundles, and builds to be non-reusable. With these changes, OSDK now avoids unnecessary network usage and reuses built objects like a standard Cargo project.

Other performance improvements include:

- ~~Makefile `TRACK_SRC_CHANGE` option: saves \~0.1s.~~ (removed)
- Using file modification times instead of sha256sum: saves ~0.2s.
- Using hard links instead of copying for most files: saves ~0.3s (only when bundle needs update).

As a result, this PR reduces the `make build` time to ~0.5s when there are no changes to the source and tests.
(For comparison: before fix: ~12.5s, offline in #1673: 1.6s-2.8s, offline + skip_build in #1673: 1.0s-2.0s).

`make ktest` is still slow due to shared base crate design. Future `--offline` option improvements could alleviate this.