@tangruize 2024-12-12
@junyang-zh I renamed `fast_copy` to `hard_link_or_copy` and removed `--source_unchanged` option. Pls review.